+++ packages/keyring-api/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- **BREAKING**: Add required `rawAmount` field to `Balance` and `FungibleAssetAmount` types for precision in balance calculations ([#426](https://github.com/MetaMask/accounts/pull/426))
 - Refine `EthAddressStruct` in order to make it compatible with the `Hex` type from `@metamask/utils` ([#405](https://github.com/MetaMask/accounts/pull/405))
 
 ## [21.3.0]